-- removed mkv support. I will resurrect it when I'll be able to completely implement it -- removed all warnings from VS2010. The only exceptions are getpid() and some constructor initialization -- enabled PCH support for VS2010. Now the building process is lightning fast! -- got rid of win32service. I will resurrect it when I will implement clustering -- some folders ignoring
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@399 01689507-d990-df11-9d7d-0017a411d416

+++ sources/applications/applestreamingclient/include/protocols/variant/messagestructure.h
@@ -97,7 +97,7 @@
 #define ASC_RES(v)						(v)["response"]
 #ifdef INCLUDE_DEBUG_LOCATIONS
 #define ASC_RES_DEBUG_FILE(v)			ASC_RES(v)["debug"]["fileName"]=__FILE__
-#define ASC_RES_DEBUG_LINE_NUMBER(v)	ASC_RES(v)["debug"]["lineNumber"]=__LINE__
+#define ASC_RES_DEBUG_LINE_NUMBER(v)	ASC_RES(v)["debug"]["lineNumber"]=(uint32_t)__LINE__
 #else
 #define ASC_RES_DEBUG_FILE(v)
 #define ASC_RES_DEBUG_LINE_NUMBER(v)
